+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 
 ### 2.2.0 Alpha 1 (XXXX-YY-ZZ)
 
+* Add `wcf\data\TDatabaseObjectOptions` and `wcf\data\TDatabaseObjectPermissions` for database object-bound options and permissions validation.
 * `wcf\system\cache\builder\EventListenerCacheBuilder` returns `wcf\data\event\listener\EventListener` objects instead of data arrays.